Sat, Sep 2, 2023, 11:22 PM +00:00

+++ UCF/COP_3223C/lecture_assignment1.c
@@ -3,8 +3,8 @@
 // Program 1
 int main(void) {
     // declaring variables, the sum, and the average
-    long int var1, var2, var3, var4, var5;
-    long int sum;
+    int var1, var2, var3, var4, var5;
+    int sum;
     double average;
 
     // reading in first input
